+++ src/components/content/resource/helpers/view-date-range.ts
@@ -3,8 +3,9 @@ import { ViewOption } from "./view-button";
 
 function dateFilter<T extends object>(field: keyof T, days: number) {
   const now = new Date();
+  // Return any that don't have a date or are within "days".
   return (data: T[]) =>
-    data.filter((d) => d[field] && differenceInDays(now, new Date(String(d[field]))) <= days);
+    data.filter((d) => !d[field] || differenceInDays(now, new Date(String(d[field]))) <= days);
 }
 
 export function getDateRangeView<T extends object>(field: Extract<keyof T, string>) {